+++ utils.py
@@ -41,7 +41,7 @@ def relu(x):
 def softmax(x, a=-1):
     """ Performs stable softmax on x, across the last axis by default """
     c = cp.exp(x-cp.amax(x, axis=a, keepdims=True))
-    return c/cp.sum(c, axis=a, keepdims=True).astype(cp.float16)
+    return c/cp.sum(c, axis=a, keepdims=True).astype(cp.float32)
 
 def apply_EI(var, ei):
     """ Applies EI masking to a square variable, according to the given
@@ -59,7 +59,7 @@ def synaptic_plasticity(h, syn_x, syn_u, constants, use_stp, hidden_size):
         syn_u = cp.minimum(1., relu(syn_u))
         h_post = syn_u*syn_x*h
     else:
-        h_post = h*cp.ones([1,1,hidden_size], dtype=cp.float16)
+        h_post = h*cp.ones([1,1,hidden_size], dtype=cp.float32)
 
     return h_post, syn_x, syn_u
 
@@ -77,7 +77,7 @@ def run_adex(V, w, I, constants):
     w_next      = adex_adaptation(V, w, constants)
     V, w, spike = adex_spike(V_next, w_next, constants)
 
-    return V.astype(cp.float16), w.astype(cp.float16), spike.astype(cp.float16)
+    return V.astype(cp.float32), w.astype(cp.float32), spike.astype(cp.float32)
 
 def adex_membrane(V, w, I, c):
     """ Calculate the new membrane potential """
@@ -112,7 +112,7 @@ def cross_entropy(mask, target, output, eps=1e-16):
     mask   = mask.astype(cp.float32)
     target = target.astype(cp.float32)
     output = output.astype(cp.float32)
-    return -cp.mean(mask[...,cp.newaxis]*target*cp.log(softmax(output)+eps), axis=(0,2,3)).astype(cp.float16)
+    return -cp.mean(mask[...,cp.newaxis]*target*cp.log(softmax(output)+eps), axis=(0,2,3)).astype(cp.float32)
 
 
 ### Optimization functions
@@ -124,7 +124,7 @@ def cross(var1, var2, rate):
 def mutate(var, num, rate, scale, epsilon=0.):
     """ Mutates a given variable by a given rate and scale,
         generating as many offspring as num """
-    mutation_mask = cp.random.random(size=[num, *var.shape], dtype=np.float32).astype(cp.float16)
+    mutation_mask = cp.random.random(size=[num, *var.shape], dtype=np.float32).astype(cp.float32)
     mutation = cp.random.normal(loc=epsilon, scale=scale, size=[num, *var.shape])
     return var[cp.newaxis,...] + mutation*mutation_mask
 
@@ -143,4 +143,4 @@ def accuracy(output, target, mask, inc_fix=False):
 
     acc = cp.sum(mask * (arg_output == arg_target), axis=(0,2))/cp.sum(mask, axis=(0,2))
 
-    return acc.astype(cp.float16)
+    return acc.astype(cp.float32)
